pyproject.toml
[tool.poetry]
name = "ahc018"
version = "0.1.0"
description = ""
authors = ["terry-u16 <terry-u16@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.2"
pandas = "^1.5.3"
matplotlib = "^3.7.0"
bayesian-optimization = "^1.4.2"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
